fix cron
